+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/type/AlarmMessage.java
@@ -35,6 +35,7 @@ public class AlarmMessage {
     private String name;
     private String message;
     private Long startTime;
+    private Long recoveryTime;
     private transient String id1;
     private final List<KeyValue> tags;
     private List<Event> events = new ArrayList<>(2);