+++ oap-server/server-storage-plugin/storage-banyandb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/banyandb/BanyanDBStorageConfig.java
@@ -100,7 +100,12 @@ public class BanyanDBStorageConfig extends ModuleConfig {
          * single request.
          */
         private int asyncProfilerTaskQueryMaxSize;
-
+        /**
+         * Max size of {@link org.apache.skywalking.oap.server.core.query.type.PprofTask} to be fetched in a
+         * single request.
+         */
+        private int pprofTaskQueryMaxSize;
+        
         private int resultWindowMaxSize = 10000;
         private int metadataQueryMaxSize = 5000;
         private int segmentQueryMaxSize = 200;