+++ docs/en/changes/changes.md
@@ -53,6 +53,7 @@
 * Self Observability: add `metrics_aggregation_queue_used_percentage` and `metrics_persistent_collection_cached_size` metrics for the OAP server.
 * Optimize metrics aggregate/persistent worker: separate `OAL` and `MAL` workers and consume pools. The dataflow signal drives the new MAL consumer, 
   the following table shows the pool size，driven mode and queue size for each worker.
+* Support pprof profiling feature. 
   
 | Worker                        | poolSize                                 | isSignalDrivenMode | queueChannelSize | queueBufferSize |
 |-------------------------------|------------------------------------------|--------------------|------------------|-----------------|